ContactPage Implementation 1
Completed preliminary work to displaying information of individual contacts on a Contact Page. Can currently display a contact's individual information page with blank sections, as well as a header for the contact, designated as the first non-null and non-empty parameter of a Contact. Implementation of a Back button was also completed.

Future work on this portion will include displaying the rest of the available information of a contact (unavailable information will not be displayed), as well as implementing the various buttons available on the page (Edit, Delete, Blacklist).