ContactPage Implementation 3
Added individual CALL and TEXT buttons next to each different phone number row (cell, home, work). Pressing these buttons will take the user to the phone or SMS app to either call or text the cell, home, or work number of the contact respectively. Started full implementation of blacklist feature, currently working on saving the blacklist status of a contact so that it remains the same when closing and restarting the app.