Minor UI fix
To remove a contact from an existing group you first have to click on it, click the dialog asking if you want to add it then click the prompt asking if you want to remove the contact b/c it already exists. If a contact already exists I changed the UI to go straight to the dialog saying that it already exists and if you want to remove it as opposed to first having to go through the dialog asking if you want to add it.